fix: handle cases where decimalSeparator is empty (#385)
* fix: handle cases where decimalSeparator is empty

Previously, for zero-decimal currencies (such as JPY), input will be lost on blur if `decimalScale` is used (no matter what value was passed in). This was down to the fact that `decimalSeparator` from `getLocaleConfig` will be `undefined` (since it doesn't exist in the output of `formatToParts`), and the fallback value of `''` was used.

This PR adds handling in the pad/trim logic to skip all processing when we see `''` as the separator.

* test: add behaviour when decimal separator is empty

* chore: just skip default value for decimal separator?

* fix: handle cases for step and key events

* test: add integration tests for 0 decimal behaviour

* test: add integration test for step

* fix: step acting weird when negatives are disallowed