fix: get group separator in locales that omit group separator for 4 digits
In some locales, when using a 4 digit number (such as '1000') the
group separator will be omited, only being added if a 5th digit
is added. This meant that the group separator was not correctly
identified for locales such as `pt-PT`.

This change updates the logic to get the locale to use more
than 4 digits, to account for those scenarios.